Update Microsoft.CodeAnalysis.Analyzers reference and fix issues it found
