+++ src/MessagePackAnalyzer/MessagePackAnalyzer.csproj
@@ -17,7 +17,7 @@
     </PackageReference>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" version="2.9.7" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" version="1.3.2" />
     <PackageReference Include="Microsoft.Composition" version="1.0.31" />
   </ItemGroup>
