+++ src/MessagePackAnalyzer/Utils/RoslynAnalyzerExtensions.cs
@@ -11,12 +11,11 @@ using System.Xml.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Formatting;
 
 namespace MessagePackAnalyzer
 {
     // Utility and Extension methods for Roslyn
-    internal static class RoslynExtensions
+    internal static class RoslynAnalyzerExtensions
     {
         public static IEnumerable<INamedTypeSymbol> GetNamedTypeSymbols(this Compilation compilation)
         {
@@ -138,31 +137,6 @@ namespace MessagePackAnalyzer
         {
             return symbol.GetMembers()
                 .Concat(symbol.AllInterfaces.SelectMany(x => x.GetMembers()));
-        }
-
-        public static CompilationUnitSyntax WithUsing(this CompilationUnitSyntax root, string name)
-        {
-            if (!root.Usings.Any(u => u.Name.ToString() == name))
-            {
-                root = root.AddUsings(SyntaxFactory.UsingDirective(SyntaxFactory.ParseName(name)).WithAdditionalAnnotations(Formatter.Annotation));
-            }
-
-            return root;
-        }
-
-        public static TNode WithFormat<TNode>(this TNode node)
-            where TNode : SyntaxNode
-        {
-            return node.WithAdditionalAnnotations(Formatter.Annotation);
-        }
-
-        public static AttributeListSyntax ParseAttributeList(string text)
-        {
-            return SyntaxFactory.ParseCompilationUnit(text)
-                .DescendantNodes()
-                .OfType<AttributeListSyntax>()
-                .First()
-                .WithFormat();
         }
     }
 }
